+++ custom_components/eventsensor/sensor.py
@@ -1,24 +1,26 @@
 """Event sensor."""
 import logging
+from typing import Any, Callable, List
 
 import voluptuous as vol
 
 from homeassistant.components.sensor import PLATFORM_SCHEMA
-from homeassistant.const import (
-    CONF_EVENT,
-    CONF_EVENT_DATA,
-    CONF_NAME,
-    CONF_STATE,
-)
+from homeassistant.config_entries import ConfigEntry, SOURCE_IMPORT
+from homeassistant.const import CONF_EVENT, CONF_EVENT_DATA, CONF_NAME, CONF_STATE
 from homeassistant.core import callback
 from homeassistant.helpers.config_validation import string
 from homeassistant.helpers.event import Event
 from homeassistant.helpers.restore_state import RestoreEntity
-from homeassistant.util import slugify
+from homeassistant.helpers.typing import (
+    ConfigType,
+    DiscoveryInfoType,
+    HomeAssistantType,
+)
+
+from .common import CONF_STATE_MAP, DOMAIN, DOMAIN_DATA, make_unique_id
 
 _LOGGER = logging.getLogger(__name__)
 
-CONF_STATE_MAP = "state_map"
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
     {
         vol.Required(CONF_NAME): string,
@@ -32,47 +34,119 @@ PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
 
 
 async def async_setup_platform(
-    hass, config, async_add_entities, discovery_info=None
+    hass: HomeAssistantType,
+    config: ConfigType,
+    async_add_entities: Callable[[List[Any], bool], None],
+    discovery_info: DiscoveryInfoType = None,
 ):
-    """Set up event sensors platform."""
-    async_add_entities(
-        [
-            EventSensor(
-                name=config.get(CONF_NAME),
-                event=config.get(CONF_EVENT),
-                event_data=config.get(CONF_EVENT_DATA, {}),
-                state=config.get(CONF_STATE),
-                state_map=config.get(CONF_STATE_MAP, {}),
+    """
+    Set up event sensors from configuration.yaml as a sensor platform.
+
+    Left just to read deprecated manual configuration.
+    """
+    if config:
+        hass.async_create_task(
+            hass.config_entries.flow.async_init(
+                DOMAIN, data=config, context={"source": SOURCE_IMPORT}
             )
-        ],
-        False,
+        )
+        _LOGGER.warning(
+            "Manual yaml config is deprecated. "
+            "You can remove it now, as it has been migrated to config entry, "
+            "handled in the Integrations menu [Sensor %s, event: %s]",
+            config.get(CONF_NAME),
+            config.get(CONF_EVENT),
+        )
+
+    return True
+
+
+async def async_setup_entry(
+    hass: HomeAssistantType,
+    config_entry: ConfigEntry,
+    async_add_entities: Callable[[List[Any], bool], None],
+):
+    """Set up the component sensors from a config entry."""
+    if DOMAIN_DATA not in hass.data:
+        hass.data[DOMAIN_DATA] = {}
+
+    _LOGGER.info(
+        f"[{config_entry.unique_id}] Entry setup with {config_entry.data} "
+        f"// {config_entry.options}"
     )
+
+    if config_entry.entry_id in hass.data[DOMAIN_DATA]:
+        _LOGGER.warning("Already in, is an update??")
+        hass.config_entries.async_update_entry(
+            config_entry, options=config_entry.options
+        )
+
+    async_add_entities([EventSensor(config_entry.data)], False)
+
+    # add an update listener to enable edition by OptionsFlow
+    if config_entry.entry_id not in hass.data[DOMAIN_DATA]:
+        hass.data[DOMAIN_DATA][
+            config_entry.entry_id
+        ] = config_entry.add_update_listener(update_listener)
+    else:
+        _LOGGER.warning("Already has a listener, is an update??")
+
+
+async def update_listener(hass: HomeAssistantType, entry: ConfigEntry):
+    """Update when config_entry options update."""
+    changes = len(entry.options) > 1 and entry.data != entry.options
+    if changes:
+        # update entry replacing data with new options, and updating unique_id and title
+        _LOGGER.critical(
+            f"ON UPDATE IN PLATFORM: {entry.data} VS {entry.options}\n"
+            f"* data id={make_unique_id(entry.data)}\n"
+            f"* opts id={make_unique_id(entry.options)}"
+        )
+        hass.config_entries.async_update_entry(
+            entry,
+            title=entry.options[CONF_NAME],
+            data=entry.options,
+            options={},
+            unique_id=make_unique_id(entry.options),
+        )
+        hass.async_create_task(hass.config_entries.async_reload(entry.entry_id))
 
 
 class EventSensor(RestoreEntity):
     """Sensor to store information originated with events."""
 
     should_poll = False
+    icon = "mdi:bullseye-arrow"
 
-    def __init__(
-        self,
-        name: str,
-        event: str,
-        event_data: dict,
-        state: str,
-        state_map: dict,
-    ):
-        self._event = event
-        self._event_data: dict = event_data
-        self._name = name
-        self._unique_id = "_".join(
-            [event, slugify(str(event_data)), state, slugify(str(state_map))]
-        )
+    def __init__(self, sensor_data: dict):
+        """Set up a new sensor mirroring some event."""
+        self._name = sensor_data.get(CONF_NAME)
+        self._event = sensor_data.get(CONF_EVENT)
+        self._state_key = sensor_data.get(CONF_STATE)
+
+        def _parse_field(raw_key: str):
+            """Enable numerical values, like press codes for remotes."""
+            try:
+                return int(raw_key)
+            except ValueError:
+                try:
+                    return float(raw_key)
+                except ValueError:
+                    return raw_key
+
+        self._event_data = {
+            _parse_field(key): _parse_field(value)
+            for key, value in sensor_data.get(CONF_EVENT_DATA, {}).items()
+        }
+        self._state_map = {
+            _parse_field(key): _parse_field(value)
+            for key, value in sensor_data.get(CONF_STATE_MAP, {}).items()
+        }
+
+        self._unique_id = make_unique_id(sensor_data)
         self._event_listener = None
         self._state = None
         self._attributes = {}
-        self._state_key = state
-        self._state_map = state_map
 
     @property
     def name(self):
@@ -123,9 +197,16 @@ class EventSensor(RestoreEntity):
         self._event_listener = self.hass.bus.async_listen(
             self._event, async_update_sensor
         )
+        _LOGGER.debug(
+            "%s: Added sensor listening to '%s' with unique_id:%s",
+            self.entity_id,
+            self._event,
+            self.unique_id,
+        )
 
     async def async_will_remove_from_hass(self):
         """Remove listeners when removing entity from Home Assistant."""
         if self._event_listener is not None:
             self._event_listener()
             self._event_listener = None
+        _LOGGER.debug("%s: Removing event listener", self.entity_id)