+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased v2.0.0](https://github.com/azogue/eventsensor/tree/feature/config-flow)
+
+[Full Changelog](https://github.com/azogue/eventsensor/compare/v1.0.0...feature/config-flow)
+
+**Implemented enhancements:**
+
+- Add config entry support + full UI support with config flows.
+
 ## [v1.0.0](https://github.com/azogue/eventsensor/tree/v1.0.0) (2020-04-17)
 
 **Initial version**